EUI-3472 post merge acrobatics
