+++ src/shared/components/case-viewer/case-full-access-view/case-full-access-view.component.spec.ts
@@ -454,6 +454,42 @@ let mockRoute: any = {
   }
 };
 
+const WORK_ALLOCATION_CASE_VIEW = {
+  case_id: CID,
+  case_type: {
+    id: CTID,
+    name: 'Test Address Book Case',
+    jurisdiction: {
+      id: JID,
+      name: 'Test',
+    },
+    printEnabled: true
+  },
+  channels: [],
+  state: {
+    id: 'CaseCreated',
+    name: 'Case created'
+  },
+  tabs: [
+    {
+      id: 'overview',
+      label: 'Overview',
+      order: 1,
+      fields: [],
+      show_condition: ''
+    },
+    {
+      id: 'caseNotes',
+      label: 'Case notes',
+      order: 2,
+      fields: [],
+      show_condition: ''
+    },
+  ],
+  triggers: TRIGGERS,
+  events: EVENTS,
+  metadataFields: METADATA,
+};
 const $DIALOG_DELETE_BUTTON = By.css('.button[title=Delete]');
 const $DIALOG_CANCEL_BUTTON = By.css('.button[title=Cancel]');
 const DIALOG_CONFIG = new MatDialogConfig();
@@ -1101,7 +1137,7 @@ xdescribe('CaseFullAccessViewComponent - print and event selector disabled', () 
   });
 });
 
-describe('CaseViewerComponent - prependedTabs', () => {
+describe('CaseFullAccessViewComponent - prependedTabs', () => {
 
   let comp: CaseFullAccessViewComponent;
   let f: ComponentFixture<CaseFullAccessViewComponent>;
@@ -1214,3 +1250,119 @@ describe('CaseViewerComponent - prependedTabs', () => {
     expect((<HTMLElement>tasksTab.querySelector('.mat-tab-label-content')).innerText).toBe('Tasks');
   });
 })
+
+// noinspection DuplicatedCode
+describe('CaseFullAccessViewComponent - Overview with prepended Tabs', () => {
+  let mockLocation: any;
+
+  let caseViewerComponent: CaseFullAccessViewComponent;
+  let componentFixture: ComponentFixture<CaseFullAccessViewComponent>;
+  let debugElement: DebugElement;
+  beforeEach((() => {
+    mockLocation = createSpyObj('location', ['path']);
+    mockLocation.path.and.returnValue('/cases/case-details/1620409659381330#caseNotes');
+    TestBed
+      .configureTestingModule({
+        imports: [
+          PaletteUtilsModule,
+          MatTabsModule,
+          ComplexModule,
+          BrowserAnimationsModule,
+          PaletteModule,
+          RouterTestingModule.withRoutes([
+            {
+              path: 'cases',
+              children: [
+                {
+                  path: 'case-details',
+                  children: [
+                    {
+                      path: ':id#overview',
+                      children: [
+                        {
+                          path: 'tasks',
+                          component: TasksContainerComponent
+                        }
+                      ]
+                    }
+                  ]
+                }
+              ]
+            }
+          ]),
+        ],
+        declarations: [
+          TasksContainerComponent,
+          CaseFullAccessViewComponent,
+          DeleteOrCancelDialogComponent,
+          // Mock
+          CaseActivityComponent,
+          EventTriggerComponent,
+          CaseHeaderComponent,
+          LinkComponent,
+          CallbackErrorsComponent,
+        ],
+        providers: [
+          FieldsUtils,
+          PlaceholderService,
+          CaseReferencePipe,
+          OrderService,
+          {
+            provide: Location,
+            useValue: mockLocation
+          },
+          ErrorNotifierService,
+          {provide: AbstractAppConfig, useClass: AppMockConfig},
+          NavigationNotifierService,
+          {provide: CaseNotifier, useValue: caseNotifier},
+          {provide: ActivatedRoute, useValue: mockRoute},
+          ActivityPollingService,
+          ActivityService,
+          HttpService,
+          HttpErrorService,
+          AuthService,
+          SessionStorageService,
+          {provide: DraftService, useValue: draftService},
+          {provide: AlertService, useValue: alertService},
+          {provide: MatDialog, useValue: dialog},
+          {provide: MatDialogRef, useValue: matDialogRef},
+          {provide: MatDialogConfig, useValue: DIALOG_CONFIG},
+          DeleteOrCancelDialogComponent
+        ]
+      })
+      .compileComponents();
+
+    componentFixture = TestBed.createComponent(CaseFullAccessViewComponent);
+    caseViewerComponent = componentFixture.componentInstance;
+    caseViewerComponent.caseDetails = WORK_ALLOCATION_CASE_VIEW;
+    caseViewerComponent.prependedTabs = [
+      {
+        id: 'tasks',
+        label: 'Tasks',
+        fields: [],
+        show_condition: null
+      },
+      {
+        id: 'roles-and-access',
+        label: 'Roles and access',
+        fields: [],
+        show_condition: null
+      }
+    ];
+    debugElement = componentFixture.debugElement;
+    componentFixture.detectChanges();
+  }));
+
+  it('should display overview tab by default', () => {
+    const matTabLabels: DebugElement = debugElement.query(By.css('.mat-tab-labels'));
+    const matTabHTMLElement: HTMLElement = matTabLabels.nativeElement as HTMLElement;
+    const tasksTab0: HTMLElement = matTabHTMLElement.children[0] as HTMLElement;
+    expect((<HTMLElement>tasksTab0.querySelector('.mat-tab-label-content')).innerText).toBe('Tasks');
+    const tasksTab1: HTMLElement = matTabHTMLElement.children[1] as HTMLElement;
+    expect((<HTMLElement>tasksTab1.querySelector('.mat-tab-label-content')).innerText).toBe('Roles and access');
+    const tasksTab2: HTMLElement = matTabHTMLElement.children[2] as HTMLElement;
+    expect((<HTMLElement>tasksTab2.querySelector('.mat-tab-label-content')).innerText).toBe('Overview');
+    const tasksTab3: HTMLElement = matTabHTMLElement.children[3] as HTMLElement;
+    expect((<HTMLElement>tasksTab3.querySelector('.mat-tab-label-content')).innerText).toBe('Case notes');
+  });
+});
