--- src/shared/components/case-viewer/case-viewer-backup/case-viewer.component.html
@@ -1,4 +1,0 @@
-<div *ngIf="isDataLoaded()">
-  <ccd-case-basic-access-view *ngIf="!hasStandardAccess()" [accessType]="userAccessType" [caseDetails]="caseDetails"></ccd-case-basic-access-view>
-  <ccd-case-full-access-view *ngIf="hasStandardAccess()" [caseDetails]="caseDetails" [hasPrint]="hasPrint" [hasEventSelector]="hasEventSelector" [prependedTabs]="prependedTabs"></ccd-case-full-access-view>
-</div>
