+++ src/shared/components/case-viewer/case-viewer.component.ts
@@ -1,168 +1,63 @@
-import { Location } from '@angular/common';
-import { AfterViewInit, Component, Input, NgZone, OnDestroy, OnInit, ViewChild, ViewContainerRef } from '@angular/core';
-import { FormControl, FormGroup } from '@angular/forms';
-import { MatDialog, MatDialogConfig, MatTabChangeEvent, MatTabGroup } from '@angular/material';
-import { ActivatedRoute, Params, Router } from '@angular/router';
-import { plainToClass } from 'class-transformer';
-import { Observable } from 'rxjs';
-import { Subject } from 'rxjs/Subject';
+import { Component, Input, NgZone, OnDestroy, OnInit } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
 import { Subscription } from 'rxjs/Subscription';
-
-import { DeleteOrCancelDialogComponent } from '../dialogs';
-import { ShowCondition } from '../../directives';
-import { Activity, CaseField, CaseTab, CaseView, CaseViewTrigger, DisplayMode, Draft, DRAFT_QUERY_PARAM, } from '../../domain';
-import {
-  ActivityPollingService,
-  AlertService,
-  DraftService,
-  ErrorNotifierService,
-  NavigationNotifierService,
-  NavigationOrigin,
-  OrderService,
-} from '../../services';
+import { AbstractAppConfig } from '../../../app.config';
+import { CaseField, CaseTab, CaseView } from '../../domain';
 import { CaseNotifier } from '../case-editor';
-import { CallbackErrorsContext } from '../error';
 
 @Component({
   selector: 'ccd-case-viewer',
-  templateUrl: './case-viewer.component.html',
-  styleUrls: ['./case-viewer.scss']
+  templateUrl: './case-viewer.component.html'
 })
-export class CaseViewerComponent implements OnInit, OnDestroy, AfterViewInit {
-  public static readonly ORIGIN_QUERY_PARAM = 'origin';
-  static readonly TRIGGER_TEXT_START = 'Go';
-  static readonly TRIGGER_TEXT_CONTINUE = 'Ignore Warning and Go';
-  static readonly UNICODE_SPACE = '%20';
-  static readonly EMPTY_SPACE = ' ';
+export class CaseViewerComponent implements OnInit, OnDestroy {
+
+  static readonly METADATA_FIELD_ACCESS_PROCEES_ID = '[ACCESS_PROCESS]';
+  static readonly NON_STANDARD_USER_ACCESS_TYPES = ['CHALLENGED', 'SPECIFIC'];
 
   @Input() public hasPrint = true;
   @Input() public hasEventSelector = true;
-  @Input() public caseDetails: CaseView;
+
   @Input() public prependedTabs: CaseTab[] = [];
-
-  public BANNER = DisplayMode.BANNER;
-  public sortedTabs: CaseTab[];
-  public caseFields: CaseField[];
-  public formGroup: FormGroup;
-  public error: any;
-  public triggerTextStart = CaseViewerComponent.TRIGGER_TEXT_START;
-  public triggerTextIgnoreWarnings = CaseViewerComponent.TRIGGER_TEXT_CONTINUE;
-  public triggerText: string = CaseViewerComponent.TRIGGER_TEXT_START;
-  public ignoreWarning = false;
-  public activitySubscription: Subscription;
+  @Input() public caseDetails: CaseView;
   public caseSubscription: Subscription;
-  public errorSubscription: Subscription;
-  public dialogConfig: MatDialogConfig;
-
-  public callbackErrorsSubject: Subject<any> = new Subject();
-  @ViewChild('tabGroup') public tabGroup: MatTabGroup;
+  public userAccessType: string;
 
   constructor(
-    private readonly ngZone: NgZone,
     private readonly route: ActivatedRoute,
-    private readonly router: Router,
-    private readonly navigationNotifierService: NavigationNotifierService,
-    private readonly orderService: OrderService,
-    private readonly activityPollingService: ActivityPollingService,
-    private readonly dialog: MatDialog,
-    private readonly alertService: AlertService,
-    private readonly draftService: DraftService,
     private readonly caseNotifier: CaseNotifier,
-    private readonly errorNotifierService: ErrorNotifierService,
-    private readonly location: Location
+    private readonly appConfig: AbstractAppConfig,
   ) {
   }
 
-  ngOnInit() {
-    this.initDialog();
-    if (!this.caseDetails) {
-      this.caseSubscription = this.caseNotifier.caseView.subscribe(caseDetails => {
-        this.caseDetails = caseDetails;
-        this.init();
-      });
-    } else {
-      this.init();
-    }
-
-    this.callbackErrorsSubject.subscribe(errorEvent => {
-      this.error = errorEvent;
-    });
-    this.errorSubscription = this.errorNotifierService.error.subscribe(error => {
-      if (error && error.status !== 401 && error.status !== 403) {
-        this.error = error;
-        this.callbackErrorsSubject.next(this.error);
-      }
-    });
+  public ngOnInit(): void {
+    this.loadCaseDetails();
   }
 
-  public isPrintEnabled(): boolean {
-    return this.caseDetails.case_type.printEnabled;
-  }
-
-  ngOnDestroy() {
-    if (this.activityPollingService.isEnabled) {
-      this.activitySubscription.unsubscribe();
-    }
-    this.callbackErrorsSubject.unsubscribe();
-    if (!this.route.snapshot.data.case) {
+  public ngOnDestroy(): void {
+    if (!!this.caseSubscription) {
       this.caseSubscription.unsubscribe();
     }
-    if (!!this.errorSubscription) {
-      this.errorSubscription.unsubscribe();
-    }
   }
 
-  public postViewActivity(): Observable<Activity[]> {
-    return this.activityPollingService.postViewActivity(this.caseDetails.case_id);
-  }
-
-  public clearErrorsAndWarnings(): void {
-    this.resetErrors();
-    this.ignoreWarning = false;
-    this.triggerText = CaseViewerComponent.TRIGGER_TEXT_START;
-  }
-
-  public applyTrigger(trigger: CaseViewTrigger): void {
-    this.error = null;
-
-    const theQueryParams: Params = {};
-
-    if (this.ignoreWarning) {
-      theQueryParams['ignoreWarning'] = this.ignoreWarning;
-    }
-
-    // we may need to take care of different triggers in the future
-    if (trigger.id === CaseViewTrigger.DELETE) {
-      const dialogRef = this.dialog.open(DeleteOrCancelDialogComponent, this.dialogConfig);
-      dialogRef.afterClosed().subscribe(result => {
-        if (result === 'Delete') {
-          this.draftService.deleteDraft(this.caseDetails.case_id)
-            .subscribe(_ => {
-              this.navigationNotifierService.announceNavigation({action: NavigationOrigin.DRAFT_DELETED});
-            }, _ => {
-              this.navigationNotifierService.announceNavigation({action: NavigationOrigin.ERROR_DELETING_DRAFT});
-            });
-        }
+  public loadCaseDetails(): void {
+    if (!this.route.snapshot.data.case) {
+      this.caseSubscription = this.caseNotifier.caseView.subscribe(caseDetails => {
+        this.caseDetails = caseDetails;
+        this.setUserAccessType(this.caseDetails);
       });
-    } else if (this.isDraft() && trigger.id !== CaseViewTrigger.DELETE) {
-      theQueryParams[DRAFT_QUERY_PARAM] = this.caseDetails.case_id;
-      theQueryParams[CaseViewerComponent.ORIGIN_QUERY_PARAM] = 'viewDraft';
-      this.navigationNotifierService.announceNavigation(
-        {
-          action: NavigationOrigin.DRAFT_RESUMED,
-          jid: this.caseDetails.case_type.jurisdiction.id,
-          ctid: this.caseDetails.case_type.id,
-          etid: trigger.id,
-          queryParams: theQueryParams
-        });
     } else {
-      this.navigationNotifierService.announceNavigation(
-        {
-          action: NavigationOrigin.EVENT_TRIGGERED,
-          queryParams: theQueryParams,
-          etid: trigger.id,
-          relativeTo: this.route
-        });
+      this.caseDetails = this.route.snapshot.data.case;
+      this.setUserAccessType(this.caseDetails);
+    }
+  }
+
+  public setUserAccessType(caseDetails: CaseView): void {
+    // remove once CCD starts sending CAM data or Access management goes live
+    this.setMockData(caseDetails);
+    if (caseDetails && Array.isArray(caseDetails.metadataFields)) {
+      const access_process = caseDetails.metadataFields.find(metadataField =>
+        metadataField.id === CaseViewerComponent.METADATA_FIELD_ACCESS_PROCEES_ID);
+      this.userAccessType = access_process ? access_process.value : null;
     }
   }
 
@@ -170,138 +65,29 @@ export class CaseViewerComponent implements OnInit, OnDestroy, AfterViewInit {
     return !!this.caseDetails;
   }
 
-  public hasTabsPresent(): boolean {
-    return this.sortedTabs.length > 0 || this.prependedTabs.length > 0;
+  public hasStandardAccess(): boolean {
+    const featureToggleOn = this.appConfig.getAccessManagementMode();
+    return featureToggleOn ? CaseViewerComponent.NON_STANDARD_USER_ACCESS_TYPES.indexOf(this.userAccessType) === -1 : true;
   }
 
-  public callbackErrorsNotify(callbackErrorsContext: CallbackErrorsContext): void {
-    this.ignoreWarning = callbackErrorsContext.ignore_warning;
-    this.triggerText = callbackErrorsContext.trigger_text;
-  }
+  // remove once Access management goes live
+  private setMockData(caseDetails: CaseView): void {
+    const accessManagementBasicViewMock = this.appConfig.getAccessManagementBasicViewMock()
+    if (accessManagementBasicViewMock.active && !caseDetails.basicFields) {
+      const access_process_index = caseDetails.metadataFields.findIndex(metadataField =>
+        metadataField.id === CaseViewerComponent.METADATA_FIELD_ACCESS_PROCEES_ID);
 
-  public isDraft(): boolean {
-    return Draft.isDraft(this.caseDetails.case_id);
-  }
-
-  public isTriggerButtonDisabled(): boolean {
-    return (this.error
-      && this.error.callbackErrors
-      && this.error.callbackErrors.length)
-      || (this.error
-        && this.error.details
-        && this.error.details.field_errors
-        && this.error.details.field_errors.length);
-  }
-
-  public ngAfterViewInit(): void {
-    let matTab;
-    const url = this.location.path(true);
-    let hashValue = url.substring(url.indexOf('#') + 1);
-    // if we have prepended tabs route to one of the prepended tabs
-    if (!url.includes('#') && this.prependedTabs && this.prependedTabs.length) {
-      const paths = url.split('/');
-      const tabName = decodeURIComponent(paths[paths.length - 1]);
-      const selectedTab: CaseTab = this.prependedTabs.find((caseTab: CaseTab) => caseTab.id.toLowerCase() === tabName.toLowerCase());
-      const tab: string = selectedTab ? selectedTab.id : 'tasks'
-      this.router.navigate(['cases', 'case-details', this.caseDetails.case_id, tab]).then(() => {
-        matTab = this.tabGroup._tabs.find((x) => x.textLabel === selectedTab.label);
-        this.tabGroup.selectedIndex = matTab.position;
-      });
-    } else {
-      const regExp = new RegExp(CaseViewerComponent.UNICODE_SPACE, 'g');
-      hashValue = hashValue.replace(regExp, CaseViewerComponent.EMPTY_SPACE);
-      matTab = this.tabGroup._tabs.find((x) =>
-        x.textLabel.replace(CaseViewerComponent.EMPTY_SPACE, '').toLowerCase() === hashValue.toLowerCase());
-      if (matTab && matTab.position) {
-        this.tabGroup.selectedIndex = matTab.position;
+      if (access_process_index > -1) {
+        caseDetails.metadataFields[access_process_index].value = accessManagementBasicViewMock.accessProcess;
+      } else {
+        const access_process: CaseField = new CaseField();
+        access_process.id = CaseViewerComponent.METADATA_FIELD_ACCESS_PROCEES_ID;
+        access_process.value = accessManagementBasicViewMock.accessProcess;
+        caseDetails.metadataFields.push(access_process);
       }
+
+      caseDetails.basicFields = accessManagementBasicViewMock.basicFields;
+
     }
   }
-
-  public tabChanged(tabChangeEvent: MatTabChangeEvent): void {
-    const tab = tabChangeEvent.tab['_viewContainerRef'] as ViewContainerRef;
-    const id = (<HTMLElement>tab.element.nativeElement).id
-    if (tabChangeEvent.index <= 1 && this.prependedTabs.length) {
-      this.router.navigate([id], {relativeTo: this.route});
-    } else {
-      this.router.navigate(['cases', 'case-details', this.caseDetails.case_id]).then(() => {
-        window.location.hash = id;
-      })
-    }
-  }
-
-  private init(): void {
-    // Clone and sort tabs array
-    this.sortedTabs = this.orderService.sort(this.caseDetails.tabs);
-    this.caseFields = this.getTabFields();
-    this.sortedTabs = this.sortTabFieldsAndFilterTabs(this.sortedTabs);
-    this.formGroup = this.buildFormGroup(this.caseFields);
-
-    if (this.activityPollingService.isEnabled) {
-      this.ngZone.runOutsideAngular(() => {
-        this.activitySubscription = this.postViewActivity().subscribe((_resolved) => {
-          // console.log('Posted VIEW activity and result is: ' + JSON.stringify(_resolved));
-        });
-      });
-    }
-
-    if (this.caseDetails.triggers && this.error) {
-      this.resetErrors();
-    }
-  }
-
-  private sortTabFieldsAndFilterTabs(tabs: CaseTab[]): CaseTab[] {
-    return tabs
-      .map(tab => Object.assign({}, tab, {fields: this.orderService.sort(tab.fields)}))
-      .filter(tab => ShowCondition.getInstance(tab.show_condition).matchByContextFields(this.caseFields));
-  }
-
-  private getTabFields(): CaseField[] {
-    const caseDataFields = this.sortedTabs.reduce((acc, tab) => {
-      return acc.concat(plainToClass(CaseField, tab.fields));
-    }, []);
-
-    return caseDataFields.concat(this.caseDetails.metadataFields);
-  }
-
-  /**
-   * For EUI-3825:
-   * Builds a FormGroup from all the CaseFields contained within the view.
-   * This FormGroup is necessary for evaluation the show/hide conditions of
-   * fields that are dependent on a field only available on a DIFFERENT tab.
-   */
-  private buildFormGroup(caseFields: CaseField[]): FormGroup {
-    let value: object = {};
-    if (caseFields) {
-      caseFields.forEach(caseField => {
-        value = {
-          ...value,
-          [caseField.id]: caseField.value
-        };
-      });
-    }
-    return new FormGroup({data: new FormControl(value)});
-  }
-
-  private initDialog(): void {
-    this.dialogConfig = new MatDialogConfig();
-    this.dialogConfig.disableClose = true;
-    this.dialogConfig.autoFocus = true;
-    this.dialogConfig.ariaLabel = 'Label';
-    this.dialogConfig.height = '245px';
-    this.dialogConfig.width = '550px';
-    this.dialogConfig.panelClass = 'dialog';
-
-    this.dialogConfig.closeOnNavigation = false;
-    this.dialogConfig.position = {
-      top: window.innerHeight / 2 - 120 + 'px', left: window.innerWidth / 2 - 275 + 'px'
-    }
-  }
-
-  private resetErrors(): void {
-    this.error = null;
-    this.callbackErrorsSubject.next(null);
-    this.alertService.clear();
-  }
-
 }
